lettings/models.py
"""
Module : models.py
Created by : Yannis Saliniere
"""

from django.core.validators import MaxValueValidator, MinLengthValidator
from django.db import models


class Address(models.Model):
    """
    Address
    """
    number = models.PositiveIntegerField(validators=[MaxValueValidator(9999)])
    street = models.CharField(max_length=64)
    city = models.CharField(max_length=64)
    state = models.CharField(max_length=2, validators=[MinLengthValidator(2)])
    zip_code = models.PositiveIntegerField(validators=[MaxValueValidator(99999)])
    country_iso_code = models.CharField(
        max_length=3, validators=[MinLengthValidator(3)]
    )

    def __str__(self):
        """
        String representation of an address instance
        """
        return f"{self.number} {self.street}"

    class Meta:
        """
        Meta class customized to pluralize the table's name
        """
        verbose_name_plural = "addresses"


class Letting(models.Model):
    """
    Letting
    """
    title = models.CharField(max_length=256)
    address = models.OneToOneField(Address, on_delete=models.CASCADE)

    def __str__(self):
        """
        String representation of a letting instance
        """
        return self.title
